+++ app/web_ui/src/routes/(app)/docs/rag_configs/[project_id]/create_rag_config/template_property_overview.svelte
@@ -1,0 +1,75 @@
+<script lang="ts">
+  import FormElement from "$lib/utils/form_element.svelte"
+  import { createEventDispatcher } from "svelte"
+  import { type RagConfigTemplate } from "../add_search_tool/rag_config_templates"
+  import PropertyList from "$lib/ui/property_list.svelte"
+
+  const dispatch = createEventDispatcher<{
+    customize_template: { template: RagConfigTemplate }
+  }>()
+
+  export let template: RagConfigTemplate
+</script>
+
+<div class="flex flex-col">
+  <FormElement
+    id="search_tool_configuration_header"
+    label="Search Configuration"
+    description="These parameters control how the search tool will extract, index, and search your documents."
+    info_description="You selected a pre-configured search tool with these parameters."
+    inputType="header_only"
+    value={null}
+  />
+  <div class="mt-2 mb-8">
+    <PropertyList
+      properties={[
+        { name: "Template Name", value: template.name },
+        {
+          name: "Extractor Model",
+          value: template.extractor.description,
+          tooltip:
+            "The model used to extract text from your documents (PDFs, images, videos, etc).",
+        },
+        {
+          name: "Chunking Strategy",
+          value: template.chunker.description,
+          tooltip:
+            "Parameters for splitting larger documents into smaller chunks for search.",
+        },
+        {
+          name: "Embedding Model",
+          value: template.embedding.description,
+          tooltip:
+            "The model used to convert text chunks into vectors for similarity search.",
+        },
+        {
+          name: "Search Index",
+          value: template.vector_store.description,
+          tooltip:
+            "How documents will be indexed and searched. Vector = semantic similarity, Full-Text = keyword search, Hybrid = both",
+        },
+        ...(template.notice_text
+          ? [
+              {
+                name: "Note",
+                value: template.notice_text,
+                warn_icon: true,
+                tooltip: template.notice_tooltip,
+              },
+            ]
+          : []),
+      ]}
+    />
+    <div class="flex flex-row items-center gap-2 mt-4">
+      <button
+        class="btn btn-sm px-6"
+        on:click={() => {
+          dispatch("customize_template", { template })
+        }}
+      >
+        Customize Configuration
+      </button>
+      <div class="badge badge-sm badge-outline">Advanced</div>
+    </div>
+  </div>
+</div>
